requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
# created-by: conda 25.1.1
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
aiohappyeyeballs=2.4.4=pypi_0
aiohttp=3.11.11=pypi_0
aiosignal=1.3.2=pypi_0
asttokens=3.0.0=pyhd8ed1ab_1
async-timeout=5.0.1=pypi_0
attrs=25.1.0=pypi_0
biopython=1.85=py310ha75aee5_1
blas=1.0=mkl
brotli-python=1.0.9=py310h6a678d5_9
bzip2=1.0.8=h5eee18b_6
ca-certificates=2025.1.31=hbcca054_0
certifi=2024.12.14=py310h06a4308_0
charset-normalizer=3.3.2=pyhd3eb1b0_0
comm=0.2.2=pyhd8ed1ab_1
cpuonly=2.0=0
cython=3.0.11=pypi_0
debugpy=1.8.11=py310h6a678d5_0
decorator=4.4.2=pypi_0
entrypoints=0.4=pyhd8ed1ab_1
exceptiongroup=1.2.2=pyhd8ed1ab_1
executing=2.1.0=pyhd8ed1ab_1
ffmpeg=4.3=hf484d3e_0
filelock=3.13.1=py310h06a4308_0
freetype=2.12.1=h4a9f257_0
frozenlist=1.5.0=pypi_0
fsspec=2024.12.0=pypi_0
gemmi=0.7.0=py310h23b7d77_0
giflib=5.2.2=h5eee18b_0
gmp=6.3.0=h6a678d5_0
gmpy2=2.2.1=py310h5eee18b_0
gnutls=3.6.15=he1e5248_0
idna=3.7=py310h06a4308_0
intel-openmp=2023.1.0=hdb19cb5_46306
ipykernel=6.29.5=pyh3099207_0
ipython=8.31.0=pyh707e725_0
jedi=0.19.2=pyhd8ed1ab_1
jinja2=3.1.5=py310h06a4308_0
joblib=1.4.2=pyhd8ed1ab_1
jpeg=9e=h5eee18b_3
jupyter_client=7.3.4=pyhd8ed1ab_0
jupyter_core=5.7.2=pyh31011fe_1
lame=3.100=h7b6447c_0
lcms2=2.16=hb9589c4_0
ld_impl_linux-64=2.40=h12ee557_0
lerc=4.0.0=h6a678d5_0
libblas=3.9.0=1_h86c2bf4_netlib
libcblas=3.9.0=8_h3b12eaf_netlib
libdeflate=1.22=h5eee18b_0
libffi=3.4.4=h6a678d5_1
libgcc=14.2.0=h77fa898_1
libgcc-ng=14.2.0=h69a702a_1
libgfortran=14.2.0=h69a702a_1
libgfortran-ng=14.2.0=h69a702a_1
libgfortran5=14.2.0=hd5240d6_1
libgomp=14.2.0=h77fa898_1
libiconv=1.16=h5eee18b_3
libidn2=2.3.4=h5eee18b_0
libjpeg-turbo=2.0.0=h9bf148f_0
liblapack=3.9.0=8_h3b12eaf_netlib
libpng=1.6.39=h5eee18b_0
libsodium=1.0.18=h36c2ea0_1
libstdcxx=14.2.0=hc0a3c3a_1
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.19.0=h5eee18b_0
libtiff=4.5.1=hffd6297_1
libunistring=0.9.10=h27cfd23_0
libuuid=1.41.5=h5eee18b_0
libwebp=1.3.2=h11a3e52_0
libwebp-base=1.3.2=h5eee18b_1
llvm-openmp=14.0.6=h9e868ea_0
lz4-c=1.9.4=h6a678d5_1
markupsafe=2.1.3=py310h5eee18b_1
matplotlib-inline=0.1.7=pyhd8ed1ab_1
mkl=2023.1.0=h213fc3f_46344
mkl-service=2.4.0=py310h5eee18b_2
mkl_fft=1.3.11=py310h5eee18b_0
mkl_random=1.2.8=py310h1128e8f_0
mpc=1.3.1=h5eee18b_0
mpfr=4.2.1=h5eee18b_0
mpmath=1.3.0=py310h06a4308_0
multidict=6.1.0=pypi_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.6.0=pyhd8ed1ab_1
nettle=3.7.3=hbbd107a_1
networkx=3.4.2=pyh267e887_2
numpy=1.24.4=pypi_0
openh264=2.1.1=h4ff587b_0
openjpeg=2.5.2=he7f1fd0_0
openssl=3.4.0=h7b32b05_1
packaging=24.2=pyhd8ed1ab_2
pandas=1.5.3=pypi_0
parso=0.8.4=pyhd8ed1ab_1
pexpect=4.9.0=pyhd8ed1ab_1
pickleshare=0.7.5=pyhd8ed1ab_1004
pillow=11.1.0=py310hcea889d_0
pip=25.0=py310h06a4308_0
platformdirs=4.3.6=pyhd8ed1ab_1
prompt-toolkit=3.0.50=pyha770c72_0
propcache=0.2.1=pypi_0
psutil=5.9.1=py310h5764c6d_0
ptyprocess=0.7.0=pyhd8ed1ab_1
pure_eval=0.2.3=pyhd8ed1ab_1
pygments=2.19.1=pyhd8ed1ab_0
pyparsing=3.2.1=pypi_0
pysocks=1.7.1=py310h06a4308_0
python=3.10.16=he870216_1
python-dateutil=2.9.0.post0=pyhff2d567_1
python_abi=3.10=2_cp310
pytorch=2.5.1=py3.10_cpu_0
pytorch-mutex=1.0=cpu
pytz=2025.1=pypi_0
pyyaml=6.0.2=py310h5eee18b_0
pyzmq=26.2.0=py310h6a678d5_0
readline=8.2=h5eee18b_0
requests=2.32.3=py310h06a4308_1
scikit-learn=1.2.2=pypi_0
scipy=1.15.1=py310hfa6ec8c_0
setuptools=75.8.0=py310h06a4308_0
six=1.17.0=pyhd8ed1ab_0
sqlite=3.45.3=h5eee18b_0
stack_data=0.6.3=pyhd8ed1ab_1
sympy=1.13.1=pypi_0
tbb=2021.8.0=hdb19cb5_0
threadpoolctl=3.5.0=pyhc1e730c_0
tk=8.6.14=h39e8969_0
torch-cluster=1.6.3=pypi_0
torch-geometric=2.6.1=pypi_0
torch-geometric-temporal=0.54.0=pypi_0
torch-scatter=2.1.2=pypi_0
torch-sparse=0.6.18=pypi_0
torch-spline-conv=1.2.2=pypi_0
torchaudio=2.5.1=py310_cpu
torchvision=0.20.1=py310_cpu
tornado=6.1=py310h5764c6d_3
tqdm=4.67.1=pypi_0
traitlets=5.14.3=pyhd8ed1ab_1
typing_extensions=4.12.2=pyha770c72_1
tzdata=2025.1=pypi_0
urllib3=2.3.0=py310h06a4308_0
wcwidth=0.2.13=pyhd8ed1ab_1
wheel=0.45.1=py310h06a4308_0
xz=5.4.6=h5eee18b_1
yaml=0.2.5=h7b6447c_0
yarl=1.18.3=pypi_0
zeromq=4.3.5=h6a678d5_0
zlib=1.2.13=h5eee18b_1
zlib-ng=2.2.3=h7955e40_0
zstd=1.5.6=hc292b87_0
